requirements.txt
slack-sdk==3.11.2
pandas==1.3.3
python-dotenv==0.19.0
pydantic==1.8.2
PyYAML==5.4.1
pytest==6.2.5
coverage==5.5
flake8==3.9.2
